GUI/requirements.txt
numpy==1.19.5
pyqt5==5.15.6
pyserial==3.5
pyqtgraph==0.11.1
pymodbus==2.5.3